chore(platform): prepare workspace packaging and secret scanner allowlist
