+++ apps/mobile/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@esparex/apps-mobile",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+    "type-check": "tsc --noEmit"
+  }
+}
