+++ core/src/domains/fraud/manifest.yaml
@@ -1,0 +1,21 @@
+id: fraud
+name: Fraud Domain
+owner: platform
+business_owner: Platform
+technical_owner: Platform
+maintainer: Fraud Squad
+maturity: stable
+visibility: public
+stability: stable
+criticality: mission-critical
+sla: tier-1
+since: "2.6"
+layer: domain
+depends_on:
+  - shared
+  - contracts
+public_api:
+  facades:
+    - FraudDetectionService
+    - SpamDetectorService
+    - DuplicateRolloutGuard
