+++ core/src/domains/trust/index.ts
@@ -1,0 +1,1 @@
+export * from './application/services/TrustService';
