+++ scripts/git/config/secret-allowlist.json
@@ -22,7 +22,9 @@
     "**/*.test.js",
     "**/*.test.jsx",
     "**/__tests__/**",
-    "**/__mocks__/**"
+    "**/__mocks__/**",
+    ".github/workflows/ci.yml",
+    "scripts/git/secret-scan.js"
   ],
   "approvedCredentials": {
     "_note": "Keys/secrets that are intentionally committed (e.g., test fixtures).",
